Change type of duration to `float` for consistency
Change type of `duration` parameter of `EasingFunction.getValue` method
and return type of `getDuration` method in `Cup` class both from `int`
to `float`, to be consistent with usages of `duration` elsewhere, as I
must have accidentally not changed all usages to be of type `float`.

+++ src/uk/co/ruben9922/cupsgame/Animation.java
@@ -40,7 +40,7 @@ class Animation {
         return change;
     }
 
-    public float getDuration() {
+    public int getDuration() {
         return duration;
     }
